==> data/DevFB/1.DownloadAndPrepareMetadata.ipynb <==
# Prepare Developing Forebrain Bulk-seq

## Data Source
- All the data downloaded from ENCODE project [here](https://www.encodeproject.org/search/?type=Experiment&status=released&lab.title=Barbara+Wold%2C+Caltech&biosample_ontology.term_name=forebrain)
- This dataset contains polyA+ RNA-seq for developing forebrain from 8 timepoints 
    - E10.5
    - E11.5
    - E12.5
    - E13.5
    - E14.5
    - E15.5
    - E16.5
    - P0
- Each time point has two replicates, a total of 16 RNA-seq
- Some experiments are pair-end sequencing, some are single end
- We start from a total of 25 fastq files from ENCODE, all experiments are single end sequencing

## Download FASTQ

# These are the URL to download files
!cat ./fastq/files.txt
# This command download all files
# !xargs -L 1 curl -O -L < files.txt

# rename the raw metadata file
!mv ?type\=Experiment\&status\=released\&files.file_type\=fastq\&lab.title\=Barbara+Wold%2C+Caltech\&biosample_ontology.term_name\=forebrain raw_metadata.csv
!ls -hl ./fastq

## Rename file and make metadata

# pandas handle table
import pandas as pd
# pathlib handle all path related
import pathlib

metadata = pd.read_csv('./fastq/raw_metadata.tsv', sep='\t', index_col=0)
print('The dataframe shape', metadata.shape)
metadata.head()

metadata.columns

### Clean metadata, select necessary columns

use_columns = [
    'Output type', 'Experiment accession', 'Biosample term id',
    'Biosample term name', 'Biological replicate(s)'
]

metadata_selected = metadata[use_columns].copy()
print('The dataframe shape', metadata_selected.shape)
metadata_selected.head()

### Concatenate timepoint information

# I manually create this sample table
exp_meta = pd.read_csv('metadata/ENCODE_experiments.tsv', sep='\t', index_col=-1)
print('The dataframe shape', exp_meta.shape)
exp_meta.head()

# These Experiment accession "ENCSR..." matched with the exp_meta, 
# our goal is add the Stage column from exp_meta to the metadata_selected
metadata_selected['Experiment accession'].head()

# single line of code can do this
metadata_selected['Stage'] = metadata_selected['Experiment accession'].map(exp_meta['Stage'])

print('The dataframe shape', metadata_selected.shape)
metadata_selected.head()

### Check some questions
- Examples to validate your metadata

print('Number of tissues', 
      metadata_selected['Biosample term name'].unique().size)
metadata_selected['Biosample term name'].unique()

print('Number of time points in each tissue')
metadata_selected.groupby('Biosample term name')['Stage'].apply(lambda i: i.unique().size)

print('Does each experiment (tissue*stage) has 2 replicates?')

# In human words, this means the number of replicates in each tissue and stage combination all equal to 2
sum(
    metadata_selected\
    .groupby(['Biosample term name', 'Stage'])\
    .apply(lambda i: i['Biological replicate(s)'].unique().size) != 2
) == 0

### Rename the files, add sample information

# create a output dir
!mkdir fastq/data

# rename the columns
rename_dict = {
    'Output type': 'count_type',
    'Experiment accession': 'experiment_id',
    'Biosample term id': 'bio_sample_id',
    'Biosample term name': 'tissue',
    'Biological replicate(s)': 'replicate',
    'Stage': 'dev_time'
}
metadata_selected.rename(columns=rename_dict, inplace=True)

# subprocess create a sub-process to execute certain shell command
import subprocess

output_path_records = []
for file_id, row in metadata_selected.iterrows():
    tissue, dev_time, rep = row[['tissue', 'dev_time', 'replicate']]
    
    # assemble the input file path and make it absolute
    input_path = pathlib.Path(f'./fastq/{file_id}.fastq.gz').absolute()
    
    # assemble the output file path
    output_path = f'./fastq/data/{tissue}_{dev_time}_{rep}_{file_id}.fastq.gz'.replace(' ', '')
    
    # save the output file path
    output_path_records.append(pathlib.Path(output_path).name)
    # and create the soft link
    subprocess.run(['ln', '-s', input_path, output_path])
    
metadata_selected['file_name'] = output_path_records
metadata_selected.head()

metadata_selected.to_csv('metadata/fastq_metadata.csv')

!ls -hl ./fastq/data/

metadata_selected
