==> data/DevFB/2.fastqTrimAndQC.ipynb <==
# FASTQ QC

import pandas as pd
import pathlib

fastq_meta = pd.read_csv('metadata/fastq_metadata.csv', index_col=0)
fastq_meta

output_dir = pathlib.Path('fastq/trimmed/').absolute()
# another way to make dir
output_dir.mkdir(exist_ok=True)

fastq_dir = pathlib.Path('fastq/data/').absolute()

## Run commands in parallel

process = 25

# this is one way to run things in parallel in python
from concurrent.futures import ProcessPoolExecutor, as_completed
import subprocess

# Here I run 25 process in parallel (on a server, not on a laptop, 
# on laptop, this can be 4 or 8 depending on your CPU)

trim_galore_path = '/home/hanliu/pkg/TrimGalore-0.6.5/trim_galore'
with ProcessPoolExecutor(process) as executor:
    #executor is a process pool, it will control the number of processes run in parallel
    
    futures = {}
    for file_name in fastq_meta['file_name']:
        file_path = fastq_dir / file_name
        command = f'{trim_galore_path} {file_path} --fastqc -o {output_dir}'
        
        # executor.submit a job (run a function with certain parameter, here the function is subprocess.run) 
        # and return a future obj. The future obj refer to this job, save it into futures first
        future = executor.submit(subprocess.run, command, shell=True, check=True)
        futures[future] = file_name
    
    # as_complete determine which future is finished
    for future in as_completed(futures):
        # get back the file name associated to this future
        file_name = futures[future]
        
        # this line is important, it check wheter the job finished without error
        # it will also got the job (subprocess.run) return, here that's not important
        _ = future.result()
        
        print(file_name, 'trim finished.')

# alternatively, this is how to run command one by one

# trim_galore_path = '/home/hanliu/pkg/TrimGalore-0.6.5/trim_galore'
# for file_name in fastq_meta['file_name']:
#     file_path = fastq_dir / file_name
#     command = f'{trim_galore_path} {file_path} --fastqc -o {output_dir}'
#     subprocess.run(command, shell=True, check=True)
#     print(file_name, 'trim finished.')

## Make metadata for trimmed fastq

# find out all the trimmed fastq, make a dict
fastq_list = list(output_dir.glob('*trimmed.fq.gz'))
fastq_list[:5]

# replace the fastq path with trimmed ones
fastq_series = pd.Series({i.name.split('_')[3]: str(i) for i in fastq_list})
fastq_meta['file_name'] = fastq_series

fastq_meta

fastq_meta.to_csv('metadata/trimmed_fastq_metadata.csv')
