==> data/DevFB/3.mapping.ipynb <==
# Salmon Mapping

## Prepare mapping commands


import pandas as pd
import pathlib

fastq_meta = pd.read_csv('./metadata/trimmed_fastq_metadata.csv', index_col=0)
fastq_meta

## Prepare salmon command for each sample

# output dir
output_dir = pathlib.Path('quant/').absolute()
output_dir.mkdir(exist_ok=True)

# set all the directories
index_dir = pathlib.Path('../ref/Salmon/salmon_index/').absolute()

# I use my own server, change this number to 4 if using laptop
# also, because salmon run in parallel internally, we just run salmon commands one by one
threads = 45

# make command for each RNA-seq sample based on the metadata
commands = {}
for (tissue, time, rep), sub_df in fastq_meta.groupby(['tissue', 'dev_time', 'replicate']):
    fastq_paths_str = ' '.join(sub_df['file_name'])
    output_name = output_dir / f'{tissue}_{time}_{rep}.quant'
    
    # assemble the final command
    command = f'salmon quant -i {index_dir} -l A -r {fastq_paths_str} --threads {threads} --validateMappings -o {output_name}'
    commands[f'{tissue}_{time}_{rep}'] = command

# a example command
command

## Run salmon

import subprocess
for name, command in commands.items():
    # once command is finished, you may want to keep a physical record, so you know its finished for sure
    # you can also use this physical to prevent rerun the command, if the execution stopped in some place
    if pathlib.Path(output_dir / name).exists():
        print('EXISTS', name)
        continue
    
    subprocess.run(command, shell=True, check=True, 
                   stdout=subprocess.PIPE, stderr=subprocess.PIPE, encoding='utf8')
    
    
    print('FINISH', name)
    with open(output_dir / name, 'w') as f:
        f.write('Oh Yeah')
    

## Clean up the flag

for name in commands.keys():
    subprocess.run(f'rm {output_dir / name}', shell=True)

## Make a metadata for salmon output

# find out all the trimmed fastq, make a dict
fastq_list = list(output_dir.glob('**/quant.sf'))
fastq_list[:5]

pd.read_csv(fastq_list[0], nrows=10, sep='\t', index_col=0)

records = []
for path in fastq_list:
    tissue, time, rep = path.parent.name.split('_')
    records.append([tissue, time, rep, str(path)])

salmon_metadata = pd.DataFrame(records, columns=['tissue', 'dev_time', 'replicate', 'salmon_count_path'])
salmon_metadata.to_csv('metadata/salmon_metadata.csv')
